Patch compiler warnings

+++ Domains.h
@@ -129,7 +129,7 @@ public:
         PlaneParam::augmentNeighborIdx(d);
 
         for (int i=0; i<3; i++)
-            for (int j=0; j<edgePoints[i].size(); j++)
+            for (size_t j=0; j<edgePoints[i].size(); j++)
                 edgePoints[i][j] += d;
     }
